requirements.txt
vosk
sounddevice
numpy
pystray
Pillow
requests
tqdm
wmi
pywin32
